* 'hebing' of e.coding.net:funenc/funenc_xa_station:
  app更新
  交接班不能接自己的班，删除只能删除草稿转台的接班
  update
  交接班更新
  update
  update
  .
  .
  。

+++ models/production_manage/production_change_shifts.py
@@ -32,16 +32,16 @@ class production_change_shifts(models.Model):
     def default_production_state(self):
 
 
-        if self.env.user.has_group('funenc_xa_station.table_station_agent'):
+        if self.env.user.has_group('funenc_xa_station.module_cstatio_nmaster'):
             # 值班站长
             return 'station_master'
-        elif self.env.user.has_group('funenc_xa_station.table_car_station_people'):
+        elif self.env.user.has_group('funenc_xa_station.module_man_on_duty'):
             # 行车
             return 'train_working'
-        elif self.env.user.has_group('funenc_xa_station.table_passenger_duty_people'):
+        elif self.env.user.has_group('funenc_xa_station.module_passenger_transport'):
             # 客运
             return 'passenger_transport'
-        elif self.env.user.has_group('funenc_xa_station.table_station_attendant'):
+        elif self.env.user.has_group('funenc_xa_station.module_depot'):
             # 站务
             return 'station_service'
         else:
@@ -50,7 +50,7 @@ class production_change_shifts(models.Model):
 
     @api.model
     def get_views(self):
-        if self.env.user.has_group('funenc_xa_station.table_station_agent'):
+        if self.env.user.has_group('funenc_xa_station.module_cstatio_nmaster'):
             list_views = self.env.ref('funenc_xa_station.funenc_xa_station_production_change_shifts_list')
             form_views = self.env.ref('funenc_xa_station.funenc_xa_station_production_change_shifts_form')
             domain = [('production_state', '=', 'station_master')]
@@ -60,7 +60,7 @@ class production_change_shifts(models.Model):
                 'form_views': form_views,
                 'domain': domain
             }
-        elif self.env.user.has_group('funenc_xa_station.table_car_station_people'):
+        elif self.env.user.has_group('funenc_xa_station.module_man_on_duty'):
             # 行车
             list_views = self.env.ref('funenc_xa_station.funenc_xa_station_production_change_shifts_list')
             form_views = self.env.ref('funenc_xa_station.production_change_train_working_shifts_form11')
@@ -71,7 +71,7 @@ class production_change_shifts(models.Model):
                 'domain': domain
             }
         # 客运
-        elif self.env.user.has_group('funenc_xa_station.table_passenger_duty_people'):
+        elif self.env.user.has_group('funenc_xa_station.module_passenger_transport'):
             list_views = self.env.ref('funenc_xa_station._production_change_shifts_list')
             form_views = self.env.ref('funenc_xa_station.station_service_form')
             domain = [('production_state', '=', 'station_service')]
@@ -216,6 +216,21 @@ class production_change_shifts(models.Model):
                                         )
 
     preparedness_state = fields.Selection(selection=[('正常', '正常'), ('异常', '异常')], default="正常", string='备品状态')  # 备品状态
+
+    is_take_over_from = fields.Integer(string='是否可接班',compute='_compute_is_take_over_from') # 用于接班按钮显示 1为显示
+
+    def _compute_is_take_over_from(self):
+        if self.env.user.id ==1:
+            for this in self:
+                this.is_take_over_from = 1
+        else:
+            ding_user_id = self.env.user.dingtalk_user.id
+            for this in self:
+                if this.state == 'change_shifts'  :
+                    if this.change_shifts_user_id.id == ding_user_id:
+                        this.is_take_over_from = 0
+                    else:
+                        this.is_take_over_from = 1
 
     @api.onchange('check_project_ids')
     def onchange_check_project_ids(self):
@@ -420,7 +435,7 @@ class production_change_shifts(models.Model):
     def create_production_change_shifts(self):
         context = dict(self.env.context or {})
 
-        if self.env.user.has_group('funenc_xa_station.table_station_agent'):
+        if self.env.user.has_group('funenc_xa_station.module_cstatio_nmaster'):
             # 值班站长
             view_form = self.env.ref('funenc_xa_station.funenc_xa_station_production_change_shifts_form').id
             return {
@@ -432,7 +447,7 @@ class production_change_shifts(models.Model):
                 'target': 'new',
                 'flags': {'initial_mode': 'edit'},
             }
-        elif self.env.user.has_group('funenc_xa_station.table_car_station_people'):
+        elif self.env.user.has_group('funenc_xa_station.module_man_on_duty'):
             # 行车
             view_form = self.env.ref('funenc_xa_station.production_change_train_working_shifts_form11').id
             return {
@@ -444,7 +459,7 @@ class production_change_shifts(models.Model):
                 'target': 'new',
                 'flags': {'initial_mode': 'edit'},
             }
-        elif self.env.user.has_group('funenc_xa_station.table_passenger_duty_people'):
+        elif self.env.user.has_group('funenc_xa_station.module_passenger_transport'):
             # 客运
             view_form = self.env.ref('funenc_xa_station.passenger_transport_train_working_shifts_form11').id
             return {
@@ -456,7 +471,7 @@ class production_change_shifts(models.Model):
                 'target': 'new',
                 'flags': {'initial_mode': 'edit'},
             }
-        elif self.env.user.has_group('funenc_xa_station.table_station_attendant'):
+        elif self.env.user.has_group('funenc_xa_station.module_depot'):
             # 站务
             view_form = self.env.ref('funenc_xa_station.station_service_form').id
             return {
@@ -497,7 +512,10 @@ class production_change_shifts(models.Model):
         }
 
     def delete(self):
-        self.unlink()
+        if self.state == 'draft':
+            self.unlink()
+        else:
+            return False
 
     def production_change_shifts_save(self):
         return {
@@ -525,18 +543,24 @@ class production_change_shifts(models.Model):
             self.state = 'take_over_from'  # 已接班
             self.take_over_from_user_id = self.env.user.dingtalk_user.id
             self.take_over_from_time = datetime.datetime.now()
+            return {
+                'name': '交接班',
+                'type': 'ir.actions.client',
+                'tag': 'change_shifts_clint',
+                'target': 'current'
+            }
 
     def get_position(self):
-        if self.env.user.has_group('funenc_xa_station.table_station_agent'):
+        if self.env.user.has_group('funenc_xa_station.module_cstatio_nmaster'):
             # 值班站长
             position = 'station_master'
-        elif self.env.user.has_group('funenc_xa_station.table_car_station_people'):
+        elif self.env.user.has_group('funenc_xa_station.module_man_on_duty'):
             # 行车
             position = 'train_working'
-        elif self.env.user.has_group('funenc_xa_station.table_passenger_duty_people'):
+        elif self.env.user.has_group('funenc_xa_station.module_passenger_transport'):
             # 客运
             position = 'passenger_transport'
-        elif self.env.user.has_group('funenc_xa_station.table_station_attendant'):
+        elif self.env.user.has_group('funenc_xa_station.module_depot'):
             # 站务
             position = 'station_service'
         else:
@@ -635,16 +659,16 @@ class production_change_shifts(models.Model):
         return obj
 
     def get_form_id(self):
-        if self.env.user.has_group('funenc_xa_station.table_station_agent'):
+        if self.env.user.has_group('funenc_xa_station.module_cstatio_nmaster'):
             # 值班站长
             form_views = self.env.ref('funenc_xa_station.funenc_xa_station_production_change_shifts_form').id
-        elif self.env.user.has_group('funenc_xa_station.table_car_station_people'):
+        elif self.env.user.has_group('funenc_xa_station.module_man_on_duty'):
             # 行车
             form_views = self.env.ref('funenc_xa_station.production_change_train_working_shifts_form11').id
-        elif self.env.user.has_group('funenc_xa_station.table_passenger_duty_people'):
+        elif self.env.user.has_group('funenc_xa_station.module_passenger_transport'):
             # 客运
             form_views = self.env.ref('funenc_xa_station.passenger_transport_train_working_shifts_form11').id
-        elif self.env.user.has_group('funenc_xa_station.table_station_attendant'):
+        elif self.env.user.has_group('funenc_xa_station.module_depot'):
             form_views = self.env.ref('funenc_xa_station.station_service_form').id
         else:
             # 票务
@@ -681,17 +705,11 @@ class production_change_shifts(models.Model):
 
     @api.model
     def handle_delete(self,id_delete):
-        # print(id_delete)
-        data = self.env['funenc_xa_station.production_change_shifts'].search([('id','=',id_delete)]).unlink()
-        return data
-
-    @api.model
-    def handle_delete_1(self,id_delete):
-        data = self.env['funenc_xa_station.production_change_shifts'].search([('id', '=', id_delete)]).unlink()
-        # print(id_delete)
-        return data
-
-
+        obj = self.search([('id', '=', id_delete)])
+        if obj.state == 'draft' or self.env.user.id ==1:
+            return obj.unlink()
+        else:
+            return False
 
 
 class put_question(models.Model):
